move serial tests in serial_test file
