+++ table/tables/index_test.go
@@ -29,9 +29,7 @@ import (
 	"github.com/pingcap/tidb/testkit"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util/codec"
-	"github.com/pingcap/tidb/util/collate"
 	"github.com/pingcap/tidb/util/mock"
-	"github.com/pingcap/tidb/util/rowcodec"
 	"github.com/stretchr/testify/require"
 )
 
@@ -89,81 +87,6 @@ func TestSingleColumnCommonHandle(t *testing.T) {
 		_, err = tablecodec.DecodeIndexKV(key, unTouchedVal, 1, tablecodec.HandleDefault,
 			tables.BuildRowcodecColInfoForIndexColumns(idx.Meta(), tblInfo))
 		require.NoError(t, err)
-	}
-}
-
-func TestMultiColumnCommonHandle(t *testing.T) {
-	collate.SetNewCollationEnabledForTest(true)
-	defer collate.SetNewCollationEnabledForTest(false)
-	tblInfo := buildTableInfo(t, "create table t (a int, b int, u varchar(64) unique, nu varchar(64), primary key (a, b), index nu (nu))")
-	var idxUnique, idxNonUnique table.Index
-	for _, idxInfo := range tblInfo.Indices {
-		idx := tables.NewIndex(tblInfo.ID, tblInfo, idxInfo)
-		if idxInfo.Name.L == "u" {
-			idxUnique = idx
-		} else if idxInfo.Name.L == "nu" {
-			idxNonUnique = idx
-		}
-	}
-	var a, b *model.ColumnInfo
-	for _, col := range tblInfo.Columns {
-		if col.Name.String() == "a" {
-			a = col
-		} else if col.Name.String() == "b" {
-			b = col
-		}
-	}
-	require.NotNil(t, a)
-	require.NotNil(t, b)
-
-	store, clean := testkit.CreateMockStore(t)
-	defer clean()
-	txn, err := store.Begin()
-	require.NoError(t, err)
-	mockCtx := mock.NewContext()
-	sc := mockCtx.GetSessionVars().StmtCtx
-	// create index for "insert t values (3, 2, "abc", "abc")
-	idxColVals := types.MakeDatums("abc")
-	handleColVals := types.MakeDatums(3, 2)
-	encodedHandle, err := codec.EncodeKey(sc, nil, handleColVals...)
-	require.NoError(t, err)
-	commonHandle, err := kv.NewCommonHandle(encodedHandle)
-	require.NoError(t, err)
-	_ = idxNonUnique
-	for _, idx := range []table.Index{idxUnique, idxNonUnique} {
-		key, _, err := idx.GenIndexKey(sc, idxColVals, commonHandle, nil)
-		require.NoError(t, err)
-		_, err = idx.Create(mockCtx, txn, idxColVals, commonHandle, nil)
-		require.NoError(t, err)
-		val, err := txn.Get(context.Background(), key)
-		require.NoError(t, err)
-		colInfo := tables.BuildRowcodecColInfoForIndexColumns(idx.Meta(), tblInfo)
-		colInfo = append(colInfo, rowcodec.ColInfo{
-			ID:         a.ID,
-			IsPKHandle: false,
-			Ft:         rowcodec.FieldTypeFromModelColumn(a),
-		})
-		colInfo = append(colInfo, rowcodec.ColInfo{
-			ID:         b.ID,
-			IsPKHandle: false,
-			Ft:         rowcodec.FieldTypeFromModelColumn(b),
-		})
-		colVals, err := tablecodec.DecodeIndexKV(key, val, 1, tablecodec.HandleDefault, colInfo)
-		require.NoError(t, err)
-		require.Len(t, colVals, 3)
-		_, d, err := codec.DecodeOne(colVals[0])
-		require.NoError(t, err)
-		require.Equal(t, "abc", d.GetString())
-		_, d, err = codec.DecodeOne(colVals[1])
-		require.NoError(t, err)
-		require.Equal(t, int64(3), d.GetInt64())
-		_, d, err = codec.DecodeOne(colVals[2])
-		require.NoError(t, err)
-		require.Equal(t, int64(2), d.GetInt64())
-		handle, err := tablecodec.DecodeIndexHandle(key, val, 1)
-		require.NoError(t, err)
-		require.False(t, handle.IsInt())
-		require.Equal(t, commonHandle.Encoded(), handle.Encoded())
 	}
 }
